+++ accounting/src/sdk/models/shared/accounttransactionlinerecordref.ts
@@ -37,24 +37,19 @@ export type AccountTransactionLineRecordRef = {
 
 /** @internal */
 export namespace AccountTransactionLineRecordRefDataType$ {
-    export const inboundSchema = z.nativeEnum(AccountTransactionLineRecordRefDataType);
-    export const outboundSchema = inboundSchema;
+    export const inboundSchema: z.ZodNativeEnum<typeof AccountTransactionLineRecordRefDataType> =
+        z.nativeEnum(AccountTransactionLineRecordRefDataType);
+    export const outboundSchema: z.ZodNativeEnum<typeof AccountTransactionLineRecordRefDataType> =
+        inboundSchema;
 }
 
 /** @internal */
 export namespace AccountTransactionLineRecordRef$ {
     export const inboundSchema: z.ZodType<AccountTransactionLineRecordRef, z.ZodTypeDef, unknown> =
-        z
-            .object({
-                dataType: AccountTransactionLineRecordRefDataType$.inboundSchema.optional(),
-                id: z.string().optional(),
-            })
-            .transform((v) => {
-                return {
-                    ...(v.dataType === undefined ? null : { dataType: v.dataType }),
-                    ...(v.id === undefined ? null : { id: v.id }),
-                };
-            });
+        z.object({
+            dataType: AccountTransactionLineRecordRefDataType$.inboundSchema.optional(),
+            id: z.string().optional(),
+        });
 
     export type Outbound = {
         dataType?: string | undefined;
@@ -65,15 +60,8 @@ export namespace AccountTransactionLineRecordRef$ {
         Outbound,
         z.ZodTypeDef,
         AccountTransactionLineRecordRef
-    > = z
-        .object({
-            dataType: AccountTransactionLineRecordRefDataType$.outboundSchema.optional(),
-            id: z.string().optional(),
-        })
-        .transform((v) => {
-            return {
-                ...(v.dataType === undefined ? null : { dataType: v.dataType }),
-                ...(v.id === undefined ? null : { id: v.id }),
-            };
-        });
+    > = z.object({
+        dataType: AccountTransactionLineRecordRefDataType$.outboundSchema.optional(),
+        id: z.string().optional(),
+    });
 }