+++ accounting/src/sdk/models/shared/accounttype.ts
@@ -18,6 +18,6 @@ export enum AccountType {
 
 /** @internal */
 export namespace AccountType$ {
-    export const inboundSchema = z.nativeEnum(AccountType);
-    export const outboundSchema = inboundSchema;
+    export const inboundSchema: z.ZodNativeEnum<typeof AccountType> = z.nativeEnum(AccountType);
+    export const outboundSchema: z.ZodNativeEnum<typeof AccountType> = inboundSchema;
 }