+++ accounting/src/sdk/models/shared/bankaccountref.ts
@@ -20,32 +20,18 @@ export type BankAccountRef = {
 
 /** @internal */
 export namespace BankAccountRef$ {
-    export const inboundSchema: z.ZodType<BankAccountRef, z.ZodTypeDef, unknown> = z
-        .object({
-            id: z.string().optional(),
-            name: z.string().optional(),
-        })
-        .transform((v) => {
-            return {
-                ...(v.id === undefined ? null : { id: v.id }),
-                ...(v.name === undefined ? null : { name: v.name }),
-            };
-        });
+    export const inboundSchema: z.ZodType<BankAccountRef, z.ZodTypeDef, unknown> = z.object({
+        id: z.string().optional(),
+        name: z.string().optional(),
+    });
 
     export type Outbound = {
         id?: string | undefined;
         name?: string | undefined;
     };
 
-    export const outboundSchema: z.ZodType<Outbound, z.ZodTypeDef, BankAccountRef> = z
-        .object({
-            id: z.string().optional(),
-            name: z.string().optional(),
-        })
-        .transform((v) => {
-            return {
-                ...(v.id === undefined ? null : { id: v.id }),
-                ...(v.name === undefined ? null : { name: v.name }),
-            };
-        });
+    export const outboundSchema: z.ZodType<Outbound, z.ZodTypeDef, BankAccountRef> = z.object({
+        id: z.string().optional(),
+        name: z.string().optional(),
+    });
 }